requirements.txt
alembic==0.8.9
bcrypt==3.1.2
certifi==2017.4.17
cffi==1.9.1
chardet==3.0.4
click==6.6
coverage==4.2
Flask==0.12
Flask-API==0.7.1
Flask-Bcrypt==0.7.1
Flask-GCM==0.2.0
Flask-HTTPAuth==3.2.2
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.6.1
gcm-client==0.1.4
gunicorn==19.7.1
idna==2.5
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.6
MarkupSafe==0.23
nose==1.3.7
psycopg2==2.6.2
pycparser==2.17
PyJWT==1.4.2
python-dateutil==2.6.0
python-editor==1.0.3
pytz==2017.2
requests==2.17.3
six==1.10.0
SQLAlchemy==1.1.4
urllib3==1.21.1
Werkzeug==0.11.13
zope.interface==4.4.2
